+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 * Replace CancelResponses by Mutation.CancelResponses on the protocol level
+* Send TextEvent before CancelResponses on text interruption
 
 ## 2022-05-01 v1.1.0