+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 * Replace CancelResponses by Mutation.CancelResponses on the protocol level
 * Send TextEvent before CancelResponses on text interruption
+* Continue audio playing after interruption in Safari
 
 ## 2022-05-01 v1.1.0