+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+## 2022-05-011 v1.2.0
+
 * Replace CancelResponses by Mutation.CancelResponses on the protocol level
 * Send TextEvent before CancelResponses on text interruption
 * Continue audio playing after interruption in Safari